Alisher

+++ src/main/java/com/example/project1/dto/LOrganizationDTO.java
@@ -13,7 +13,7 @@ public class LOrganizationDTO {
     private String address;
     private Long managerID;
     private String urlImage;
-    private Integer Rate;
+    private Double Rate;
     private Integer users;
     private Long generalOrganizationID;
     private Set<Long> categoryID;
